fix for int64 on posts

+++ internal/database/posts.sql.go
@@ -262,7 +262,9 @@ SELECT
     s.network AS network,
     r.likes,
     r.reposts,
-    COALESCE(r.likes, 0) + COALESCE(r.reposts, 0) AS interactions,
+    (
+        COALESCE(r.likes, 0) + COALESCE(r.reposts, 0)
+    )::bigint AS interactions,
     r.views
 FROM
     posts p
@@ -290,7 +292,7 @@ type GetRecentPostsForUserRow struct {
 	Network           sql.NullString
 	Likes             sql.NullInt64
 	Reposts           sql.NullInt64
-	Interactions      int32
+	Interactions      int
 	Views             sql.NullInt64
 }
